Changed a lof of things
